+++ package.json
@@ -39,6 +39,7 @@
     "@mdx-js/loader": "^1.5.7",
     "@mdx-js/react": "^1.5.7",
     "@next/mdx": "^9.3.1",
+    "babel-plugin-import-glob-array": "^0.2.0",
     "comma-number": "2.0.1",
     "date-fns": "^2.11.0",
     "emotion-theming": "^10.0.27",
