+++ pages/blog/resume.mdx
@@ -1,5 +1,5 @@
 ---
-title: 'You Don't Need A Two-Page Resume'
+title: "You Don't Need A Two-Page Resume"
 publishedAt: '2020-01-14'
 summary: 'Want a better programming resume? Limit it to a single page. See how my resume has transformed over the past five years with some actionable advice.'
 image: '/static/images/resume/banner.jpg'
